Event/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.2
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
phonenumbers==8.13.0
phonenumberslite==8.13.0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==1.4.2
zope.interface==5.5.2
